test with 500mb long to support more than 2gb files

+++ app/src/main/java/com/owncloud/android/utils/EncryptionUtils.java
@@ -629,7 +629,7 @@ public final class EncryptionUtils {
 
             // check authentication tag
             byte[] extractedAuthenticationTag = new byte[AUTHENTICATION_TAG_LENGTH];
-            inputStream.getChannel().position((int) inputFile.length() - AUTHENTICATION_TAG_LENGTH);
+            inputStream.getChannel().position(inputFile.length() - AUTHENTICATION_TAG_LENGTH);
 
             int readBytes = inputStream.read(extractedAuthenticationTag,
                                              0,
